requirements.txt
Django==1.10
djangorestframework==3.7.3
drf-nested-routers==0.90.0